clean code清理：表达式中不要混用有符号数和无符号数
# message auto-generated for no-merge-commit merge:
merge clean-code into master

clean code清理：表达式中不要混用有符号数和无符号数  

Created-by: msprofDeveloper

Match-id-47a38415f9b9f30f57843673bad526ca8430cb9a

+++ analysis/csrc/domain/data_process/ai_task/msproftx_device_processor.h
@@ -19,7 +19,7 @@
 namespace Analysis {
 namespace Domain {
 // model_id, index_id, stream_id, task_id, timestamp
-using TxDeviceData = std::tuple<uint32_t, uint32_t, uint32_t, uint32_t, double>;
+using TxDeviceData = std::tuple<uint32_t, uint32_t, uint32_t, uint32_t, uint64_t>;
 using OriMsprofTxDeviceData = std::vector<TxDeviceData>;
 
 class MsprofTxDeviceProcessor : public DataProcessor {
